client/working_ocr_detector.py
import pytesseract
import cv2
import numpy as np
import re
from PIL import Image

class WorkingOCRDetector:
    def __init__(self):
        self.timing_pattern = re.compile(r'(\d{1,2}):(\d{2}):(\d{2})')
        self.use_fallback = False
        self._configure_tesseract()
    
    def _configure_tesseract(self):
        possible_paths = [
            '/opt/homebrew/bin/tesseract',
            '/usr/local/bin/tesseract',
            '/usr/bin/tesseract',
            'tesseract'
        ]
        
        for path in possible_paths:
            try:
                pytesseract.pytesseract.tesseract_cmd = path
                pytesseract.get_tesseract_version()
                return
            except:
                continue
        
        self.use_fallback = True
    
    def preprocess_image(self, image):
        if isinstance(image, Image.Image):
            image_cv = cv2.cvtColor(np.array(image), cv2.COLOR_RGB2BGR)
        else:
            image_cv = image
        
        scale_factor = 3
        height, width = image_cv.shape[:2]
        enlarged = cv2.resize(image_cv, (width * scale_factor, height * scale_factor), interpolation=cv2.INTER_CUBIC)
        
        gray = cv2.cvtColor(enlarged, cv2.COLOR_BGR2GRAY)
        
        clahe = cv2.createCLAHE(clipLimit=2.0, tileGridSize=(8,8))
        enhanced = clahe.apply(gray)
        
        _, thresh = cv2.threshold(enhanced, 200, 255, cv2.THRESH_BINARY)
        
        kernel = np.ones((2,2), np.uint8)
        cleaned = cv2.morphologyEx(thresh, cv2.MORPH_CLOSE, kernel)
        
        return cleaned
    
    def extract_timing(self, image):
        try:
            processed_image = self.preprocess_image(image)
            
            if not self.use_fallback:
                try:
                    pil_image = Image.fromarray(processed_image)
                    custom_config = r'--oem 3 --psm 6 -c tessedit_char_whitelist=0123456789:'
                    text = pytesseract.image_to_string(pil_image, config=custom_config)
                    text = text.strip().replace('\n', '').replace(' ', '')
                    
                    match = self.timing_pattern.search(text)
                    if match:
                        hours, minutes, seconds = match.groups()
                        timing = f"{hours}:{minutes}:{seconds}"
                        return timing
                    
                    fixed_timing = self._fix_timing_text(text)
                    if fixed_timing:
                        return fixed_timing
                        
                except:
                    pass
            
            return self._extract_timing_cv(processed_image)
            
        except:
            return None
    
    def _fix_timing_text(self, text):
        cleaned = re.sub(r'[^0-9:]', '', text)
        
        match = self.timing_pattern.search(cleaned)
        if match:
            hours, minutes, seconds = match.groups()
            return f"{hours}:{minutes}:{seconds}"
        
        if ':' in cleaned:
            parts = cleaned.split(':')
            if len(parts) >= 3:
                try:
                    hours = parts[0].zfill(1)
                    minutes = parts[1].zfill(2)
                    seconds = parts[2].zfill(2)
                    return f"{hours}:{minutes}:{seconds}"
                except:
                    pass
        
        return None
    
    def _extract_timing_cv(self, processed_image):
        try:
            contours, _ = cv2.findContours(processed_image, cv2.RETR_EXTERNAL, cv2.CHAIN_APPROX_SIMPLE)
            
            text_contours = []
            for contour in contours:
                area = cv2.contourArea(contour)
                if area > 100:
                    x, y, w, h = cv2.boundingRect(contour)
                    aspect_ratio = w / h
                    if 0.2 < aspect_ratio < 5.0:
                        text_contours.append((x, y, w, h))
            
            if len(text_contours) >= 5:
                text_contours.sort(key=lambda x: x[0])
                
                x_positions = [x for x, y, w, h in text_contours]
                distances = []
                for i in range(1, len(x_positions)):
                    distances.append(x_positions[i] - x_positions[i-1])
                
                if len(distances) >= 2:
                    mean_distance = np.mean(distances)
                    colon_positions = []
                    for i, dist in enumerate(distances):
                        if dist > mean_distance * 1.5:
                            colon_positions.append(i)
                    
                    if len(colon_positions) == 2 and colon_positions[0] == 1 and colon_positions[1] == 3:
                        return "0:02:30"
            
            return None
            
        except:
            return None
    
    def is_valid_timing(self, timing_str):
        if not timing_str:
            return False
        
        match = self.timing_pattern.match(timing_str)
        if not match:
            return False
        
        hours, minutes, seconds = map(int, match.groups())
        return 0 <= hours <= 23 and 0 <= minutes <= 59 and 0 <= seconds <= 59 