client/requirements.txt
pyautogui==0.9.54
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.21.0
pytesseract>=0.3.10 